+++ src/factory/project.ts
@@ -1,3 +1,4 @@
+import { IService as IPaymentService } from './service/paymentService';
 import SortType from './sortType';
 
 /**
@@ -36,6 +37,7 @@ export interface ISettings {
     importEventsInWeeks?: number;
     onEventChanged?: IOnEventChanged;
     onReservationStatusChanged?: IOnReservationStatusChanged;
+    paymentServices?: IPaymentService[];
 }
 
 /**